debian: Remove r4d configuration
Elbe does not start any daemons at target installalation time.
The r4d configuration does not succeed without a rd4 daemon start
because the database is created on the first run since r4d v1.6.

As a temporary workaround, remove the r4dcfg calls and the service
restarts with it but keep the r4d Pre-Depends because package users
might transitively depend on that.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Benedikt Spranger <b.spranger@linutronix.de>